+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.8.5 (2021-03-03)
+
+### Bug fixes
+
+* Avoid JNI crash in leaveBreadcrumb by pushing local frame
+  [#214](https://github.com/bugsnag/bugsnag-unity/pull/214)
+
+* Respect autoNotify flag on Android
+  [#207](https://github.com/bugsnag/bugsnag-unity/pull/207)
+
 ## 4.8.4 (2020-10-05)
 
 ### Enhancements